Refactor branch
I'm experimenting with refactoring the codebase of reenforcements to be more centralized, which includes renaming a lot of files. For the sake of not sabotaging the mod for people casually using it, this'll be it's own branch until 1:1 parity is made with the new system. So far mechanoids are loaded, and the other enemy classes are awaiting the event handler being finished (I'm going with a data driven approach that will use a populated table instead of various hard coded values across many handlers).