Mimic Cans (partial)
Changed more about the spawn handler. I think the best idea going forward might be to make it so classes do their configs for firemode/traits/etc in postbeginplay() per class, instead of being touched by the handler (since the latter means there would need to be a generic namespace of config values it can push to. ...Not impossible, actually, may be worth doing that with an array of structs per class that's usually empty).