Add mimic barrels/cans, rework cvars
Adds mimic barrels and cans to the current refactor, and uses varying strategies to track cvars at points in time (throwaways for spawn cvars, functions that constantly check for behavioral cvars, etc). Most logic is deferred from the spawners for now, but this info can probably be set via WorldThingSpawned() for each monster class if the performance hit is too high.